Match the order and number of tiles we had as default in N
The fact that location is not set as a default tile but is turned on
seems a bit sneaky. I don't have a problem with the tile being turned
on but it should be visible to the user so if they wish to turn it off,
they can.

Also the order never made sense. Since Lollipop, maybe Kitkat the cell tile
has always come after WiFi.

Change-Id: Ib1c6eb86e702caff3cadcebab8b0f0e8441a6610
Signed-off-by: RahifM <faizel326@gmail.com>